# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: Forum API
  description: |
    API for managing forums, topics, and posts. This API enables users to create, read, update, and delete forums along with their associated topics and posts.

    ## Forum Features
    - **Forums**: Manage various discussion forums.
    - **Topics**: Create topics within each forum to facilitate focused discussions.
    - **Posts**: Users can create and manage posts within topics.
    - **User Interaction**: Users can comment on posts and engage with one another.

    ## contact
    - **Name**: Suchakhri Dangthaisong
    - **Email**: chakhree.h@gmail.com
    - **Git**: https://github.com/Chakhrxx

  version: '1.0'
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: 'API key needed to access the endpoints.'
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT needed to access secured endpoints.'
...
security:
  - ApiKeyAuth: []
  - JwtAuth: []
...
paths:
...
  /api/users:
    post:
      tags:
        - Users
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
              required:
                - username
                - email
              example:
                username: 'Smith57'
                email: 'Smith57@test.com'
      responses:
        '201':
          description: Successfully created user
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  email:
                    type: string
                    format: email
                  id:
                    type: string
                    format: uuid
                  created:
                    type: string
                    format: date-time
                  updated:
                    type: string
                    format: date-time
                example:
                  username: 'Smith57'
                  email: 'Smith57@test.com'
                  id: '75d07e3b-9b64-4822-89bf-777c97b9e51b'
                  created: '2024-10-26T13:33:12.348Z'
                  updated: '2024-10-26T13:33:12.348Z'

    get:
      tags:
        - Users
      summary: Retrieve all users
      security:
        - JwtAuth: []
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    username:
                      type: string
                    email:
                      type: string
                      format: email
                    id:
                      type: string
                      format: uuid
                    created:
                      type: string
                      format: date-time
                    updated:
                      type: string
                      format: date-time
              example:
                - username: 'Smith57'
                  email: 'Smith57@test.com'
                  id: '75d07e3b-9b64-4822-89bf-777c97b9e51b'
                  created: '2024-10-26T13:33:12.348Z'
                  updated: '2024-10-26T13:33:12.348Z'
                - username: 'DoeJohn'
                  email: 'DoeJohn@test.com'
                  id: 'c47d8f5e-99de-4c29-9e05-4bb5cf7e1e70'
                  created: '2024-10-25T11:22:33.123Z'
                  updated: '2024-10-25T11:22:33.123Z'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Retrieve a user by ID
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to retrieve
          schema:
            type: string
            example: '75d07e3b-9b64-4822-89bf-777c97b9e51b'
      responses:
        '200':
          description: Successfully retrieved user
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  email:
                    type: string
                    format: email
                  id:
                    type: string
                    format: uuid
                  created:
                    type: string
                    format: date-time
                  updated:
                    type: string
                    format: date-time
                example:
                  username: 'Smith57'
                  email: 'Smith57@test.com'
                  id: '75d07e3b-9b64-4822-89bf-777c97b9e51b'
                  created: '2024-10-26T13:33:12.348Z'
                  updated: '2024-10-26T13:33:12.348Z'
        '404':
          description: User not found

    patch:
      tags:
        - Users
      summary: Update a user by ID
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to update
          schema:
            type: string
            example: '75d07e3b-9b64-4822-89bf-777c97b9e51b'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
              example:
                username: 'Smith57'
                email: 'Smith57@test.com'
      responses:
        '200':
          description: Successfully updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  email:
                    type: string
                    format: email
                  id:
                    type: string
                    format: uuid
                  created:
                    type: string
                    format: date-time
                  updated:
                    type: string
                    format: date-time
                example:
                  username: 'Smith57'
                  email: 'Smith57@test.com'
                  id: '75d07e3b-9b64-4822-89bf-777c97b9e51b'
                  created: '2024-10-26T13:33:12.348Z'
                  updated: '2024-10-26T13:33:12.348Z'

    delete:
      tags:
        - Users
      summary: Delete a user by ID
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to delete
          schema:
            type: string
            example: '75d07e3b-9b64-4822-89bf-777c97b9e51b'
      responses:
        '204':
          description: Successfully deleted user
        '404':
          description: User not found
...